Camera ray to world seems to be working fine
